+++ src/main/java/com/dis/bot/model/age/AgeEffect.java
@@ -1,4 +1,4 @@
-package com.dis.bot.model;
+package com.dis.bot.model.age;
 
 import lombok.Data;
 import org.springframework.data.annotation.Id;
@@ -6,11 +6,13 @@ import org.springframework.data.mongodb.core.mapping.Document;
 
 @Document("age_effects")
 @Data
-public class AgeSkill {
+public class AgeEffect {
 
     @Id
     private String id;
     private String characterId;
     private String name;
-    private Long points;
+    private Long duration;
+    private Boolean active;
+    private Long startRound;
 }
