+++ src/main/java/com/dis/bot/model/age/AgeEquipment.java
@@ -1,17 +1,17 @@
-package com.dis.bot.model;
+package com.dis.bot.model.age;
 
 import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-@Document("age_powers")
+@Document("age_equipments")
 @Data
-public class AgePower {
+public class AgeEquipment {
 
     @Id
     private String id;
     private String characterId;
     private String name;
-    private Long level;
     private String description;
+    private Boolean isMagic;
 }
