+++ src/main/java/com/dis/bot/model/age/AgeFeat.java
@@ -1,17 +1,23 @@
-package com.dis.bot.model;
+package com.dis.bot.model.age;
 
 import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-@Document("age_equipments")
+@Document("age_feats")
 @Data
-public class AgeEquipment {
+public class AgeFeat {
 
     @Id
     private String id;
     private String characterId;
     private String name;
-    private String description;
-    private Boolean isMagic;
+    private FeatType type;
+
+    public enum FeatType {
+        ADVENTURER,
+        CHAMPION,
+        EPIC
+    }
+
 }
