+++ src/main/java/com/dis/bot/model/age/AgeSkill.java
@@ -1,17 +1,16 @@
-package com.dis.bot.model;
+package com.dis.bot.model.age;
 
 import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-@Document("age_spells")
+@Document("age_effects")
 @Data
-public class AgeSpell {
+public class AgeSkill {
 
     @Id
     private String id;
     private String characterId;
     private String name;
-    private Long level;
-    private String description;
+    private Long points;
 }
