+++ src/main/java/com/dis/bot/model/age/AgeSpell.java
@@ -1,23 +1,17 @@
-package com.dis.bot.model;
+package com.dis.bot.model.age;
 
 import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-@Document("age_feats")
+@Document("age_spells")
 @Data
-public class AgeFeat {
+public class AgeSpell {
 
     @Id
     private String id;
     private String characterId;
     private String name;
-    private FeatType type;
-
-    public enum FeatType {
-        ADVENTURER,
-        CHAMPION,
-        EPIC
-    }
-
+    private Long level;
+    private String description;
 }
